Fix build error caused by unknown header/status/method reference
The problem causes by HTTPDomain.resolve_xref method which tries to create the
right reference node for header/status/method. However, on http_statuscode_role
or http_header_role call it has no knowledge about the inliner (this is docutils
internal rst parser thing) so it pass it as None causing the issue. But the
inliner knows the reporter to write out about the problem.

In final solution, we have to get reporter thought resolving doctree by it name
and replace functions for reference build with XRefRole instances to simulate
similar behaviour, interface and features access for both cases when
header/status/method are defined as role and when as directive fields.

By default, strict mode is turned on, so in case of usage custom headers and
response statuses you'll see build errors in the output. This decision made
with thoughts about quality control in mind. If it annoys you much, set
`http_strict_mode = False` in your conf.py.